Consolidate Testem Setup
In preperation for adding more complexity I've modularized our testem
setup to make it a bit easier to read and to add new screenshots and
preferences to.
