+++ .github/workflows/ci.yml
@@ -166,7 +166,7 @@ jobs:
         run: pnpm install
       - name: Capture Screenshots
         if: steps.screenshot-cache.outputs.cache-hit != 'true'
-        run: pnpm --filter frontend exec ember exam --parallel=3 --load-balance --config-file=testem.screenshots.js
+        run: pnpm --filter frontend exec ember exam --parallel=3 --load-balance --config-file=testem/screenshots-config.js
       - name: Rename screenshots
         if: steps.screenshot-cache.outputs.cache-hit != 'true'
         run: |
