+++ packages/frontend/eslint.config.mjs
@@ -86,8 +86,8 @@ export default [
       '**/*.cjs',
       'config/**/*.js',
       'tests/dummy/config/**/*.js',
+      'testem/*.js',
       'testem.js',
-      'testem*.js',
       'index.js',
       '.prettierrc.js',
       '.stylelintrc.js',
